+++ packages/core_media_data/lib/core_media_data.dart
@@ -4,3 +4,4 @@ export "src/media_database_benchmark_models.dart";
 export "src/media_graph.dart";
 export "src/media_ranking.dart";
 export "src/profile_data_ownership_models.dart";
+export "src/similar_media_resolver.dart";